requirements.txt
beautifulsoup4==4.12.3
certifi==2024.8.30
charset-normalizer==3.4.0
confluent-kafka==2.6.1
idna==3.10
linkedin-api==2.3.1
lxml==5.3.0
python-dotenv==1.0.1
requests==2.32.3
soupsieve==2.6
urllib3==2.2.3
